Add header file for C-code #95
Add header file for C-code #95
